move JsonRpcResponseResult and JsonRpcStreamedResult to separate classes
